volatility surface dupire
Starting from the dupire relations between quoted market prices, exercise prices and option lifetimes we return:
a. a matrix of implied volatilities corresponding to their quoted market prices. 
b. a 2D interpolator
c. a new matrix of requested volatilities given the interpolator g and a matrix of strike prices and time-to-maturities.